python/requirements.txt
tensorflow==2.7.2
joblib==1.0.1
pandas==1.2.4
sklearn
pytorch==1.10.2
pytorch-lightning==1.5.10